Fixes diagram deps cycle
