+++ src/components/diagrams/singleLineDiagram/position-diagram-pane.tsx
@@ -11,11 +11,11 @@ import { useSelector } from 'react-redux';
 import { PARAM_LANGUAGE, PARAM_USE_NAME } from '../../../utils/config-params';
 import PositionDiagram from './position-diagram';
 import { SLD_DISPLAY_MODE } from '../../network/constants';
-import { DiagramType } from '../diagram-common';
 import { getVoltageLevelSingleLineDiagram } from '../../../services/study/network';
 import { FC } from 'react';
 import { AppState } from 'redux/reducer';
 import { UUID } from 'crypto';
+import { DiagramType } from '../diagram.type';
 
 interface PositionDiagramPaneProps {
     open: boolean;
