+++ src/components/diagrams/singleLineDiagram/single-line-diagram-content.tsx
@@ -9,7 +9,6 @@ import { useCallback, useLayoutEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { RunningStatus } from '../../utils/running-status';
 import {
-    DiagramType,
     getEquipmentTypeFromFeederType,
     MAX_HEIGHT_SUBSTATION,
     MAX_HEIGHT_VOLTAGE_LEVEL,
@@ -18,7 +17,6 @@ import {
     MIN_HEIGHT,
     MIN_WIDTH,
     styles,
-    useDiagram,
 } from '../diagram-common';
 import withEquipmentMenu from '../../menus/equipment-menu';
 import BaseEquipmentMenu, { MapEquipment } from '../../menus/base-equipment-menu';
@@ -53,6 +51,8 @@ import { AppState } from 'redux/reducer';
 import { UUID } from 'crypto';
 import { INVALID_LOADFLOW_OPACITY } from '../../../utils/colors';
 import { useParameterState } from 'components/dialogs/parameters/use-parameters-state';
+import { DiagramType } from '../diagram.type';
+import { useDiagram } from '../use-diagram';
 
 type EquipmentMenuState = {
     position: [number, number];
