+++ src/components/study-pane.jsx
@@ -11,12 +11,13 @@ import PropTypes from 'prop-types';
 import { ReportViewerTab } from './report-viewer-tab';
 import { ResultViewTab } from './result-view-tab';
 import TabPanelLazy from './results/common/tab-panel-lazy';
-import { DiagramType, useDiagram } from './diagrams/diagram-common';
 import { isNodeBuilt } from './graph/util/model-functions';
 import { TableWrapper } from './spreadsheet/table-wrapper';
 import ParametersTabs from './parameters-tabs';
 import MapViewer from './map-viewer';
 import { StudyView } from './utils/utils';
+import { DiagramType } from './diagrams/diagram.type.js';
+import { useDiagram } from './diagrams/use-diagram.js';
 
 const styles = {
     map: {
